Set GOMAXPROCS to at least number of keys in pool.
Some (most?) PKCS#11 modules will block the thread they are on during
operations. This means that if you're running pkcs11key on a machine that has
fewer CPUs (and thus a smaller GOMAXPROCS) than your HSM, you will not be able
to utilize the full performance of your HSM, unless you increase GOMAXPROCS.
This change handles that for you automatically.