API change: use pubkey rather than privkey label. (#10)
Previously, a Key was created with four arguments:
module, token label, private key label, and pin. However, most users don't
necessarily know the label for their private key, and it's relatively complex to
find out. Additionally, the old interface made it easy to construct a Signer
using a private key that didn't match the public key you were planning to sign
with. If such a signer was used to sign something, everything would appear
to work, but the signature would be invalid.

In this API-breaking change, I switch out the privateKeyLabel argument for a
crypto.PublicKey. This matches common usage in that the calling code usually
knows the PublicKey it wishes to use, because it's present in a certificate.

This has the significant added benefit that we can delete a lot of code that was
dedicated to reading out the public key structures from the token and turning
them into a `crypto.PublicKey` object to satisfy the Signer interface.

In this change I also add a findObject utility function that wraps the
FindObjectsInint / FindObjects / FindObjectsFinal trio of calls in pkcs11.

Replaces #5. cc @titanous

If we merge this, I'll give it a release tag of v2.0.0 to indicate the breaking change.