Improve documentation.
Link to the latest version of the PKCS#11 standard, now managed by OASIS
(and on HTTPS instead of FTP).

Reference the module path with major version from the old-style path
documentation.

Give an example of how to benchmark with a Yubikey.

Fix command line flags in comment in pkcs11key_bench_test.go.