Switch to GitHub Actions workflow for test (#32)
Also:
 - Fix tests
 - Remove .travis.yml
 - Remove fallback to SoftHSMv1 for Travis

The benchmark was using the public key from the certificate template, which was bogus. Switching it to use the public key from the `Pool` fixes things.

The latest version of SoftHSMv2 assigns a random slot ID on creation rather than accepting the one it's given, so parse out the assigned slot ID from its output.